backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vladislav-76 <sillmaril2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-resized = "^1.0.2"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
